+++ examples/asa/scripts/transfer/contract-owned-asa/contract-to-bob.js
@@ -26,7 +26,7 @@ async function run (runtimeEnv, deployer) {
   const txGroup = [
     // Stateful call
     {
-      type: types.TransactionType.CallNoOpSSC,
+      type: types.TransactionType.CallApp,
       sign: types.SignType.SecretKey,
       fromAccount: bob,
       appID: appInfo.appID,