+++ examples/bond/scripts/run/common/common.js
@@ -45,7 +45,15 @@ async function fundAccount (deployer, accountAddress) {
   await executeTransaction(deployer, algoTxnParams);
 }
 
-// Opt-In lsigs to a given asa
+/**
+ * Returns Opt-In lsigs to a given asa transaction
+ * Note: only manager is allowed to opt-in lsigs to asa
+ * This is because, no other user can spam asa opt-in in lsigs
+ * @param deployer deployer object
+ * @param managerAcc manager account
+ * @param lsig logic signature
+ * @param assetIndex asset index
+ */
 async function optInTx (deployer, managerAcc, lsig, assetIndex) {
   const optInTx = [
     {
@@ -68,6 +76,209 @@ async function optInTx (deployer, managerAcc, lsig, assetIndex) {
   await executeTransaction(deployer, optInTx);
 }
 
+/**
+ * Creates a group transaction for buying bond tokens.
+ * @param buyer buyer account
+ * @param issuerLsig Bond issuer logic signature
+ * @param amount amount of bonds
+ * @param algoAmount amount of Algo to pay for the bonds
+ * @param appID Bond application index
+ * @param bondID Bond ASA index
+ */
+function buyTx (buyer, issuerLsig, amount, algoAmount, appID, bondID) {
+  return [
+    // Algo transfer from buyer to issuer
+    {
+      type: types.TransactionType.TransferAlgo,
+      sign: types.SignType.SecretKey,
+      fromAccount: buyer,
+      toAccountAddr: issuerLsig.address(),
+      amountMicroAlgos: algoAmount,
+      payFlags: { totalFee: 2000 }
+    },
+    // Bond token transfer from issuer's address
+    {
+      type: types.TransactionType.TransferAsset,
+      sign: types.SignType.LogicSignature,
+      fromAccountAddr: issuerLsig.address(),
+      lsig: issuerLsig,
+      toAccountAddr: buyer.addr,
+      amount: amount,
+      assetID: bondID,
+      payFlags: { totalFee: 0 }
+    },
+    // call to bond-dapp
+    {
+      type: types.TransactionType.CallApp,
+      sign: types.SignType.SecretKey,
+      fromAccount: buyer,
+      appID: appID,
+      payFlags: { totalFee: 1000 },
+      appArgs: ['str:buy']
+    }
+  ];
+}
+
+/**
+ * Returns issue group transaction
+ * @param creatorAccount creator acccount
+ * @param issuerLsig Bond issuer logic signature
+ * @param appID Bond application index
+ * @param bondID Bond ASA index
+ */
+function issueTx (creatorAccount, issuerLsig, appID, bondID) {
+  return [
+    // Bond asa transfer to issuer's address
+    {
+      type: types.TransactionType.TransferAsset,
+      sign: types.SignType.SecretKey,
+      fromAccount: creatorAccount,
+      toAccountAddr: issuerLsig.address(),
+      amount: 1e6,
+      assetID: bondID,
+      payFlags: { }
+    },
+    // call to bond-dapp
+    {
+      type: types.TransactionType.CallApp,
+      sign: types.SignType.SecretKey,
+      fromAccount: creatorAccount,
+      appID: appID,
+      payFlags: {},
+      appArgs: ['str:issue']
+    }
+  ];
+}
+
+/**
+ * Returns create dex transaction
+ * @param managerAccount App manager account
+ * @param appID Bond application index
+ * @param issuerLsig Bond issuer logic signature
+ * @param dexLsig Bond dex logic signature
+ * @param creatorAccount Bond ASA creator account
+ * @param unsoldBonds balance of oldBond tokens hold by the issuer (amount of unsold bonds)
+ * @param newBondID new bond ASA index
+ * @param oldBondID old bond ASA index
+ * @param total total number of bonds issued to buyers
+ * @param couponValue value of coupon
+ */
+function createDexTx (
+  managerAccount, appID, issuerLsig,
+  dexLsig, creatorAccount, unsoldBonds,
+  newBondID, oldBondID, total, couponValue
+) {
+  return [
+    // call to bond-dapp
+    {
+      type: types.TransactionType.CallApp,
+      sign: types.SignType.SecretKey,
+      fromAccount: managerAccount,
+      appID: appID,
+      payFlags: {},
+      appArgs: ['str:create_dex'],
+      accounts: [issuerLsig.address(), dexLsig.address()]
+    },
+    // New bond token transfer to issuer's address
+    {
+      type: types.TransactionType.TransferAsset,
+      sign: types.SignType.SecretKey,
+      fromAccount: creatorAccount,
+      toAccountAddr: issuerLsig.address(),
+      amount: unsoldBonds,
+      assetID: newBondID,
+      payFlags: { totalFee: 1000 }
+    },
+    // burn tokens
+    {
+      type: types.TransactionType.TransferAsset,
+      sign: types.SignType.LogicSignature,
+      fromAccountAddr: issuerLsig.address(),
+      lsig: issuerLsig,
+      toAccountAddr: creatorAccount.addr,
+      amount: unsoldBonds,
+      assetID: oldBondID,
+      payFlags: { totalFee: 1000 }
+    },
+    // Transfer app.total amount of new Bonds to dex lsig
+    {
+      type: types.TransactionType.TransferAsset,
+      sign: types.SignType.SecretKey,
+      fromAccount: creatorAccount,
+      toAccountAddr: dexLsig.address(),
+      amount: total,
+      assetID: newBondID,
+      payFlags: { totalFee: 1000 }
+    },
+    // Algo transfer to dex address
+    {
+      type: types.TransactionType.TransferAlgo,
+      sign: types.SignType.SecretKey,
+      fromAccount: creatorAccount,
+      toAccountAddr: dexLsig.address(),
+      amountMicroAlgos: Number(total) * Number(couponValue),
+      payFlags: { totalFee: 1000 }
+    }
+  ];
+}
+
+/**
+ * Returns redeem transaction
+ * @param buyerAccount buyer account
+ * @param dexLsig Bond dex logic signature
+ * @param amount amount of bonds to be redeemed
+ * @param oldBondID old bond token index
+ * @param newBondID new bond token index
+ * @param couponValue value of coupon
+ * @param appID Bond application index
+ */
+function redeemCouponTx (
+  buyerAccount, dexLsig, amount, oldBondID,
+  newBondID, couponValue, appID
+) {
+  return [
+    // Transfer tokens to dex lsig.
+    {
+      type: types.TransactionType.TransferAsset,
+      sign: types.SignType.SecretKey,
+      fromAccount: buyerAccount,
+      toAccountAddr: dexLsig.address(),
+      amount: amount,
+      assetID: oldBondID,
+      payFlags: { totalFee: 3000 }
+    },
+    // New bond token transfer to buyer's address
+    {
+      type: types.TransactionType.TransferAsset,
+      sign: types.SignType.LogicSignature,
+      fromAccountAddr: dexLsig.address(),
+      lsig: dexLsig,
+      toAccountAddr: buyerAccount.addr,
+      amount: amount,
+      assetID: newBondID,
+      payFlags: { totalFee: 0 }
+    },
+    {
+      type: types.TransactionType.TransferAlgo,
+      sign: types.SignType.LogicSignature,
+      fromAccountAddr: dexLsig.address(),
+      lsig: dexLsig,
+      toAccountAddr: buyerAccount.addr,
+      amountMicroAlgos: Number(amount) * Number(couponValue),
+      payFlags: { totalFee: 0 }
+    },
+    // call to bond-dapp
+    {
+      type: types.TransactionType.CallApp,
+      sign: types.SignType.SecretKey,
+      fromAccount: buyerAccount,
+      appID: appID,
+      payFlags: { totalFee: 1000 },
+      appArgs: ['str:redeem_coupon']
+    }
+  ];
+}
+
 module.exports = {
   issuePrice,
   asaDef,
@@ -76,5 +287,9 @@ module.exports = {
   tokenMap,
   couponValue,
   optInTx,
-  nominalPrice
+  nominalPrice,
+  buyTx,
+  issueTx,
+  createDexTx,
+  redeemCouponTx
 };