+++ packages/runtime/src/lib/constants.ts
@@ -25,6 +25,10 @@ export const LogicSigMaxCost = 20000;
 export const MaxAppProgramCost = 700;
 export const LogicSigMaxSize = 1000;
 export const MaxAppProgramLen = 1024;
+export const ALGORAND_MAX_APP_ARGS_LEN = 16;
+export const ALGORAND_MAX_TX_ACCOUNTS_LEN = 4;
+// the assets and application arrays combined and totaled with the accounts array can not exceed 8
+export const ALGORAND_MAX_TX_ARRAY_LEN = 8;
 
 export const MAX_ALGORAND_ACCOUNT_ASSETS = 1000;
 export const MAX_ALGORAND_ACCOUNT_APPS = 10;